+++ packages/theme-github/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## bytemd-theme-github@0.0.4
+
+> 2026-07-31
+
+- Upgrade to `@tofrankie/github-markdown-css@0.0.5`
+
 ## bytemd-theme-github@0.0.3
 
 > 2026-07-30